+++ visula/src/init_wgpu.rs
@@ -104,8 +104,8 @@ pub async fn init(proxy: EventLoopProxy<CustomEvent>, window: Window) {
     });
 
     let platform = Platform::new(PlatformDescriptor {
-        physical_width: size.width as u32,
-        physical_height: size.height as u32,
+        physical_width: size.width,
+        physical_height: size.height,
         scale_factor: window.scale_factor(),
         font_definitions: FontDefinitions::default(),
         style: Default::default(),
